# sdk-server/docs/swagger.yaml
definitions:
  base.PageResp:
    properties:
      currentPage:
        description: 当前第几页
        type: integer
      list:
        description: 数据列表
      pageSize:
        description: 分页大小
        type: integer
      total:
        description: 总条数
        type: integer
    type: object
  base.ReqId:
    properties:
      id:
        description: 主键ID
        type: integer
    type: object
  base.ReqIds:
    properties:
      ids:
        description: 多id
        items:
          type: integer
        type: array
    type: object
  base.Resp:
    properties:
      code:
        description: 返回码
        type: integer
      data:
        description: 数据
      msg:
        description: 消息
        type: string
      reqId:
        description: '`json:"请求id"`'
        type: string
    type: object
...
  dto.NoticeDto:
    properties:
      count:
        type: integer
      key:
        type: string
      list:
        items:
          $ref: '#/definitions/dto.NoticeItem'
        type: array
      name:
        type: string
      total:
        type: integer
    type: object
  dto.NoticeItem:
    properties:
      beginAt:
        type: integer
      content:
        type: string
      createdAt:
        type: integer
      endAt:
        type: integer
      id:
        type: integer
      noticeType:
        type: integer
      op:
        type: integer
      opId:
        type: integer
      reminderStatus:
        type: integer
      reminderTime:
        type: integer
      status:
        type: integer
      title:
        type: string
      type:
        type: integer
    type: object
...
  dto.SysDeptDto:
    properties:
      deptId:
        description: 主键
        type: integer
      deptName:
        type: string
      deptPath:
        type: string
      email:
        type: string
      id:
        description: 主键
        type: integer
      leader:
        type: string
      name:
        description: 部门名称
        type: string
      parentId:
        type: integer
      phone:
        type: string
      principal:
        description: 部门领导
        type: string
      remark:
        description: 备注
        type: string
      sort:
        type: integer
      status:
        type: integer
      teamId:
        description: 团队id
        type: integer
      type:
        description: 状态
        type: integer
    type: object
  dto.SysDeptGetPageReq:
    properties:
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      status:
        type: integer
      teamId:
        description: 团队id
        type: integer
    type: object
...
  dto.TaskDto:
    properties:
      beginAt:
        description: 开始时间
        type: string
      content:
        description: 任务内容
        type: string
      endAt:
        description: 结束时间
        type: string
      id:
        description: 主键
        type: integer
      op:
        description: 操作类型
        type: integer
      opId:
        description: 操作id
        type: integer
      reminderStatus:
        description: 提醒状态 1开启 2关闭
        type: integer
      reminderTime:
        description: 提醒时间
        type: string
      status:
        description: 状态1开启2关闭
        type: integer
      taskType:
        description: 任务类型
        type: integer
      teamId:
        description: 团队id
        type: integer
      title:
        description: 任务标题
        type: string
      userId:
        description: 用户id
        type: integer
    type: object
  dto.TaskGetPageReq:
    properties:
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      status:
        description: 状态1开启2关闭ReminderStatus int `json:"reminderStatus" query:"column:reminder_status"`
          //提醒状态 1开启 2关闭
        type: integer
      teamId:
        type: integer
      userId:
        type: integer
    type: object
...
  models.SysDept:
    properties:
      createBy:
        description: 创建者
        type: integer
      createdAt:
        description: 创建时间
        type: string
      deptPath:
        description: 路径
        type: string
      email:
        description: 邮箱
        type: string
      id:
        description: 主键
        type: integer
      name:
        description: 部门名称
        type: string
      parentId:
        description: 上级部门
        type: integer
      phone:
        description: 手机号
        type: string
      principal:
        description: 部门领导
        type: string
      remark:
        description: 备注
        type: string
      sort:
        description: 排序
        type: integer
      status:
        description: 状态
        type: integer
      teamId:
        description: 团队id
        type: integer
      type:
        description: 状态
        type: integer
      updateBy:
        description: 更新者
        type: integer
      updatedAt:
        description: 最后更新时间
        type: string
    type: object
...
  models.Task:
    properties:
      beginAt:
        description: 开始时间
        type: string
      content:
        description: 任务内容
        type: string
      createdAt:
        description: 创建时间
        type: string
      endAt:
        description: 结束时间
        type: string
      id:
        description: 主键
        type: integer
      op:
        description: 操作类型
        type: integer
      opId:
        description: 操作id
        type: integer
      reminderStatus:
        description: 提醒状态 1开启 2关闭
        type: integer
      reminderTime:
        description: 提醒时间
        type: string
      status:
        description: 状态1开启2关闭
        type: integer
      taskType:
        description: 任务类型
        type: integer
      teamId:
        description: 团队id
        type: integer
      title:
        description: 任务标题
        type: string
      updatedAt:
        description: 更新时间
        type: string
      userId:
        description: 用户id
        type: integer
    type: object
...
info:
  contact:
    email: baowk1024@gmail.com
    name: baowk
    url: https://github.com/baowk/dilu
  description: 致力于做一个开发快速，运行稳定的框架
  title: Dilu API
  version: V0.0.1
paths:
...
  /api/team/dept/all:
    post:
      consumes:
      - application/json
      parameters:
      - description: 团队id
        in: header
        name: teamId
        type: integer
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysDeptGetPageReq'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.SysDept'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: 获取全部部门
      tags:
      - team-Dept
  /api/team/dept/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: 团队id
        in: header
        name: teamId
        type: integer
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysDeptDto'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.SysDept'
              type: object
      security:
      - Bearer: []
      summary: 创建SysDept
      tags:
      - team-Dept
  /api/team/dept/del:
    post:
      consumes:
      - application/json
      parameters:
      - description: 团队id
        in: header
        name: teamId
        type: integer
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/base.ReqIds'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.SysDept'
              type: object
      security:
      - Bearer: []
      summary: 删除SysDept
      tags:
      - team-Dept
  /api/team/dept/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: 团队id
        in: header
        name: teamId
        type: integer
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/base.ReqId'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.SysDept'
              type: object
      security:
      - Bearer: []
      summary: 获取SysDept
      tags:
      - team-Dept
  /api/team/dept/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 团队id
        in: header
        name: teamId
        type: integer
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SysDeptDto'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.SysDept'
              type: object
      security:
      - Bearer: []
      summary: 更新SysDept
      tags:
      - team-Dept
...
  /api/team/task/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: 团队id
        in: header
        name: teamId
        type: integer
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.TaskDto'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.Task'
              type: object
      security:
      - Bearer: []
      summary: 创建Task
      tags:
      - notice-Task
  /api/team/task/del:
    post:
      consumes:
      - application/json
      parameters:
      - description: 团队id
        in: header
        name: teamId
        type: integer
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/base.ReqIds'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.Task'
              type: object
      security:
      - Bearer: []
      summary: 删除Task
      tags:
      - notice-Task
  /api/team/task/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: 团队id
        in: header
        name: teamId
        type: integer
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/base.ReqId'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.Task'
              type: object
      security:
      - Bearer: []
      summary: 获取Task
      tags:
      - notice-Task
  /api/team/task/my:
    post:
      consumes:
      - application/json
      parameters:
      - description: 团队id
        in: header
        name: teamId
        type: integer
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.TaskGetPageReq'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/dto.NoticeDto'
              type: object
      security:
      - Bearer: []
      summary: 获取Task列表
      tags:
      - notice-Task
  /api/team/task/page:
    post:
      consumes:
      - application/json
      parameters:
      - description: 团队id
        in: header
        name: teamId
        type: integer
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.TaskGetPageReq'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/base.PageResp'
                  - properties:
                      list:
                        items:
                          $ref: '#/definitions/models.Task'
                        type: array
                    type: object
              type: object
      security:
      - Bearer: []
      summary: 获取Task列表
      tags:
      - notice-Task
  /api/team/task/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 团队id
        in: header
        name: teamId
        type: integer
      - description: body
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.TaskDto'
      responses:
        "200":
          description: '{"code": 200, "data": [...]}'
          schema:
            allOf:
            - $ref: '#/definitions/base.Resp'
            - properties:
                data:
                  $ref: '#/definitions/models.Task'
              type: object
      security:
      - Bearer: []
      summary: 更新Task
      tags:
      - notice-Task
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"